Replace TryLoadTable with LoadTable.
TryLoadTable just returns a true or a false based on whether it was able to load a table.
Due to this, we can run in NPE if a table was  not loaded and we try to run operations on it.
Replacing it with LoadTable because this throws an exception if the table isn't loaded. We can
catch this exception and display a relevent error message.